+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix: #4 Incorrect excerpt on new lines
 - Fix: Sometimes the first image was not used for the preview
+- Add: Additional variables added, `{{url}}`, `{{createdTime}}` and `{{updatedTime}}` #6
 
 ## v0.2.0 (2023-12-28)